refactor: split vessel_drafter_scene.py into focused modules
Fixes #19
